# openapi.yaml
openapi: 3.1.0
info:
  title: GlassBox API
  version: 0.1.0
  description: |
    Contract for the GlassBox FastAPI gateway. `/api/*` routes are a
    1:1 port of the legacy Flask dashboard
    (backend-source/dashboard/DASHBOARD_PRO.py). `/admin/*` routes are
    new (Phase 4, Agent Factory) and require an admin JWT.

    Live signals are also available over WebSocket at `/ws/signals`
    (see the `x-websocket` extension below) -- OpenAPI 3.1 has no
    native operation type for WS, so it's documented rather than
    declared as a path.

    Every response carries an `X-Request-ID` header (a fresh UUID4 per
    request), also echoed in that request's structured server-side log
    line -- quote it when reporting an issue against a specific call.
servers:
  - url: http://localhost:8000
security: []
paths:
...
  /admin/agents:
    get:
      operationId: listAgents
      summary: List agent configs
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AgentConfig" }
    post:
      operationId: createAgent
      summary: Create an agent config
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentConfig" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentConfig" }
  /admin/agents/{agentId}:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getAgent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AgentConfig" } } } }
        "404": { description: Not found }
    put:
      operationId: updateAgent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentConfig" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AgentConfig" } } } }
        "404": { description: Not found }
    delete:
      operationId: deleteAgent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
        "404": { description: Not found }
  /admin/orchestrations:
    get:
      operationId: listOrchestrations
      summary: List orchestration configs
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrchestrationConfig" }
    post:
      operationId: createOrchestration
      summary: Create an orchestration config
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrchestrationConfig" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrchestrationConfig" }
  /admin/orchestrations/{orchestrationId}:
    parameters:
      - name: orchestrationId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getOrchestration
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OrchestrationConfig" } } } }
        "404": { description: Not found }
    put:
      operationId: updateOrchestration
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrchestrationConfig" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OrchestrationConfig" } } } }
        "404": { description: Not found }
    delete:
      operationId: deleteOrchestration
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
        "404": { description: Not found }
  /admin/agents/{agentId}/test-run:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: testRunAgent
      summary: Run one agent against freeform input as a background job
      description: |
        Deterministic agents never touch an LLM provider (input is
        treated as a ticker symbol and resolved against the existing
        quant signal engine); LLM agents call the configured provider
        with the agent's system prompt/params. Always runs as a
        background job -- see /ws/jobs/{jobId} for live progress.
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentTestRunRequest" }
      responses:
        "202":
          description: Accepted -- job created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobAccepted" }
        "404": { description: Agent not found }
  /admin/orchestrations/{orchestrationId}/run:
    parameters:
      - name: orchestrationId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: runOrchestration
      summary: Run an orchestration (sequential/parallel/committee_vote) as a background job
      description: |
        committee_vote reduces through a deterministic confidence-weighted
        vote over participating agents' outputs -- never an LLM judge.
        Each agent is bounded by agent_timeout_s; the whole run is bounded
        by run_budget_s. Agents that miss their budget are marked
        degraded and excluded rather than failing the run.
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrchestrationRunRequest" }
      responses:
        "202":
          description: Accepted -- job created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobAccepted" }
        "404": { description: Orchestration not found }
  /admin/jobs/{jobId}:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getJob
      summary: Poll a background job's status (fallback to /ws/jobs/{jobId})
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/JobStatus" } } } }
        "404": { description: Not found }
...
  /admin/audit-log:
    get:
      operationId: getAuditLog
      summary: Paginated history of every admin mutation and login attempt
      description: |
        Every agent/orchestration create-update-delete, test-run,
        orchestration-run, report.generate, and login attempt (success or
        failure -- failures record the attempted username, never the
        password) is written here at the time it happens. Newest first.
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedAuditLog" }
  /admin/llm-calls:
    get:
      operationId: getLLMCalls
      summary: Paginated cost/latency log of every LLM call made anywhere in the app
      description: |
        Covers both agent test-runs / orchestration runs and
        /reports/generate narration -- every path that calls an
        LLMProvider writes a row here. Newest first.
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedLLMCalls" }
  /reports/generate:
    post:
      operationId: generateReport
      summary: Generate an LLM-narrated daily report as a background job
      description: |
        Reads the latest reports/daily_*.json plus current live signals,
        builds a narration prompt, and calls the selected provider.
        Result (a DailyReportNarrative) is delivered via /ws/jobs/{jobId}
        or polled via /admin/jobs/{jobId}, and persisted -- see
        /reports/narratives.
      tags: [reports]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportGenerateRequest" }
      responses:
        "202":
          description: Accepted -- job created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobAccepted" }
        "404": { description: No report data available to narrate }
  /reports/narratives:
    get:
      operationId: listReportNarratives
      summary: List generated report narratives
      tags: [reports]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DailyReportNarrative" }
  /reports/narratives/{narrativeId}:
    parameters:
      - name: narrativeId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getReportNarrative
      summary: Fetch one generated narrative (viewer-accessible, no admin gate)
      tags: [reports]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/DailyReportNarrative" } } } }
        "404": { description: Not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AgentParams:
      type: object
      properties:
        temperature: { type: number, default: 0.7 }
        top_p: { type: number, default: 1.0 }
        max_tokens: { type: integer, default: 1024 }
        extra: { type: object, additionalProperties: true }
    AgentConfig:
      type: object
      required: [name, role, type]
      properties:
        id: { type: string, readOnly: true }
        name: { type: string }
        role: { type: string, example: "Macro Analyst" }
        type: { type: string, enum: [deterministic, llm] }
        provider: { type: string, enum: [vllm, ollama, gemini, claude], nullable: true }
        model: { type: string, nullable: true }
        params: { $ref: "#/components/schemas/AgentParams" }
        system_prompt: { type: string, nullable: true }
        tools:
          type: array
          items: { type: string }
        enabled: { type: boolean, default: true }
    OrchestrationConfig:
      type: object
      required: [name, mode, agent_ids]
      properties:
        id: { type: string, readOnly: true }
        name: { type: string }
        mode: { type: string, enum: [sequential, parallel, committee_vote] }
        agent_ids:
          type: array
          items: { type: string }
        coordinator: { type: string, default: vn_engine }
        schedule: { type: string, nullable: true, example: "0 8 * * 1-5" }
        agent_timeout_s: { type: number, default: 30.0, description: "Per-agent wall-clock budget for one run" }
        run_budget_s: { type: number, default: 120.0, description: "Whole-orchestration wall-clock budget" }
...
    LLMCallLog:
      type: object
      required: [id, provider, model, started_at, status]
      properties:
        id: { type: string }
        agent_id: { type: string, nullable: true }
        provider: { type: string, enum: [vllm, ollama, gemini, claude] }
        model: { type: string }
        started_at: { type: string, format: date-time }
        duration_ms: { type: integer, nullable: true }
        tokens_in: { type: integer, nullable: true }
        tokens_out: { type: integer, nullable: true }
        status: { type: string, enum: [running, ok, error, timeout] }
        error: { type: string, nullable: true }
        estimated_cost_usd:
          type: number
          nullable: true
          description: >-
            Null when the provider/model has no pricing entry (e.g. Gemini
            with no GEMINI_*_COST_PER_1M env vars set). Always 0.0 for
            self-hosted ollama/vllm. Token counts feeding this estimate are
            approximate (whitespace-based), so treat the figure as
            directional, not billing-grade.
    PaginatedLLMCalls:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/LLMCallLog" } }
        total: { type: integer }
    AuditLogEntry:
      type: object
      required: [id, actor, action, resource_type, detail, created_at]
      properties:
        id: { type: string }
        actor: { type: string, description: "Username, or \"unknown\" for a failed login attempt" }
        action: { type: string, example: "agent.create" }
        resource_type: { type: string, enum: [agent, orchestration, report, auth] }
        resource_id: { type: string, nullable: true }
        detail: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
    PaginatedAuditLog:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AuditLogEntry" } }
        total: { type: integer }
    JobStatus:
      type: object
      required: [job_id, kind, status, created_at]
      properties:
        job_id: { type: string }
        kind: { type: string, enum: [agent_test_run, orchestration_run, report_generate] }
        status: { type: string, enum: [queued, running, done, error] }
        created_at: { type: string, format: date-time }
        result: { type: object, additionalProperties: true, nullable: true }
        error: { type: string, nullable: true }
    JobAccepted:
      type: object
      required: [job_id]
      properties:
        job_id: { type: string }
    AgentTestRunRequest:
      type: object
      required: [input]
      properties:
        input: { type: string, description: "Freeform prompt/context; treated as a ticker symbol for deterministic agents" }
    OrchestrationRunRequest:
      type: object
      properties:
        input: { type: string, nullable: true }
    ReportGenerateRequest:
      type: object
      required: [model]
      properties:
        date: { type: string, nullable: true, description: "YYYYMMDD; defaults to today if omitted (report data itself is always the latest on disk)" }
        provider: { type: string, enum: [vllm, ollama, gemini, claude], default: ollama }
        model: { type: string }
        agent_id: { type: string, nullable: true, description: "Reuse an existing agent's system_prompt for narration style" }
    DailyReportNarrative:
      type: object
      required: [id, date, provider, model, narrative, created_at]
      properties:
        id: { type: string }
        date: { type: string }
        provider: { type: string, enum: [vllm, ollama, gemini, claude] }
        model: { type: string }
        narrative: { type: string }
        created_at: { type: string, format: date-time }
...